+++ src/container/App.js
@@ -1,11 +1,11 @@
 import 'css-reset-and-normalize';
 import '../styles/style.css';
-import Projects from '../pages/Projects';
+import ProjectPage from '../pages/Project';
 // import Tasks from '../pages/Task';
 
 const App = () => {
   const main = document.createElement('main');
-  main.innerHTML = Projects();
+  main.innerHTML = ProjectPage();
   // main.append(Projects());
   // main.append(Tasks);
